--- a/master_changes.md
+++ b/master_changes.md
@@ -50,6 +50,7 @@ users)
 ## Show
   * Add `depexts` to default printer [#4898 @rjbou]
   * Make `opam show --list-files <pkg>` fail with not found when `<pkg>` is not installed [#4956 @kit-ty-kate - fix #4930]
+  * Improve performance of opam show by 300% when the package to show is given explicitly or unique [#4998 @kit-ty-kate - fix #4997 and partially #4172]
 
 ## Var
   *