Fixes #4997 and partially https://github.com/ocaml/opam/issues/4172

Before:
```
opam show --raw dune.2.9.0  0.94s user 0.32s system 97% cpu 1.286 total
```
After:
```
./opam show --raw dune.2.9.0  0.37s user 0.05s system 97% cpu 0.433 total
```