@kit-ty-kate 2022-01-15
Some timings for `opam show --raw dune.2.9.0`:
* opam 2.0.10: 0.7s
* opam 2.1.2: 3.0s
* this PR: 0.6s :tada:

(timings taken via an archlinux docker image)