@AltGr 2022-01-17
Haha, nice one! I guess the costly part was the `Lazy.force st.available_packages` ?